+++ assets/app/templates/editor/edit-file.html
