+++ assets/app/templates/site/list-item.html
