+++ assets/app/views/editor/blocks/heading.js
