+++ assets/app/views/editor/edit-list.js
@@ -3,7 +3,7 @@ var fs = require('fs');
 var Backbone = require('backbone');
 var _ = require('underscore');
 
-var templateHtml = fs.readFileSync(__dirname + '/../templates/AuthenticateTemplate.html').toString();
+var templateHtml = fs.readFileSync(__dirname + '../../templates/authenticate.html').toString();
 
 var EditorFileListView = Backbone.View.extend({
   tagName: 'ul',
