+++ assets/app/views/editor/edit-main.js
