+++ assets/app/views/site/add.js
@@ -5,10 +5,10 @@ var _ = require('underscore');
 
 /* this require looks gross but it enables us to have one place
 to set template data that is available to the front and back end apps */
-var SiteTemplates = require('../../../config/templates').templates;
+var SiteTemplates = require('../../../../config/templates').templates;
 
-var SiteModel = require('../models/Site').model;
-var templateHtml = fs.readFileSync(__dirname + '/../templates/AddSiteTemplate.html').toString();
+var SiteModel = require('../../models/Site').model;
+var templateHtml = fs.readFileSync(__dirname + '/../../templates/site/add.html').toString();
 
 var AddSiteView = Backbone.View.extend({
   tagName: 'div',
